Entity and App config implementation - Spring

+++ backend/habitStats_Server/src/main/java/pte/mik/habitstatsserver/dto/ActionStatDto.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 @Data
 public class ActionStatDto implements Serializable {
-    private final Integer id;
+    private final Long id;
     private final String title;
-    private final Integer categoryId;
-    private final Integer unit_typeId;
+    private final Long categoryId;
+    private final Long unit_typeId;
 }
