+++ backend/habitStats_Server/src/main/java/pte/mik/habitstatsserver/dto/MapperGoalDto.java
@@ -7,6 +7,6 @@ import java.io.Serializable;
 @Data
 public class MapperGoalDto implements Serializable {
     // TODO Make it use lists for ease of use
-    private final Integer statId;
-    private final Integer goalId;
+    private final Long statId;
+    private final Long goalId;
 }
