+++ backend/habitStats_Server/src/main/java/pte/mik/habitstatsserver/entity/stat/ActionProgressDto.java
@@ -3,13 +3,12 @@ package pte.mik.habitstatsserver.entity.stat;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.sql.Timestamp;
 import java.time.Instant;
 
 @Data
 public class ActionProgressDto implements Serializable {
-    private final Integer id;
-    private final Integer statId;
+    private final Long id;
+    private final Long statId;
     private final Float value;
     private final Instant timestamp;
 }
