+++ backend/habitStats_Server/src/main/java/pte/mik/habitstatsserver/repository/stat/ProgressRepository.java
@@ -9,7 +9,7 @@ import pte.mik.habitstatsserver.entity.stat.Progress;
 import java.util.List;
 
 @Repository
-public interface ProgressRepository extends JpaRepository<Progress, Integer>{
+public interface ProgressRepository extends JpaRepository<Progress, Long>{
     @Query("select p FROM Progress p WHERE p.stat.id = ?1")
-    public List<Progress> getByStatId(Integer statId);
+    public List<Progress> getByStatId(Long statId);
 }
