+++ backend/habitStats_Server/src/main/java/pte/mik/habitstatsserver/repository/stat/UnitTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pte.mik.habitstatsserver.entity.stat.UnitType;
 
 @Repository
-public interface UnitTypeRepository extends JpaRepository<UnitType, Integer> {
+public interface UnitTypeRepository extends JpaRepository<UnitType, Long> {
 }
